ci(c/driver_manager,c/driver/postgres): add macOS wheels
